C - Bit manipulation. Task 5. 101. also fixing data types for previous tasks

+++ bit_manipulation/3-set_bit.c
@@ -11,7 +11,7 @@
  */
 int set_bit(unsigned long int *n, unsigned int index)
 {
-	unsigned int power = 1;
+	unsigned long int power = 1;
 	unsigned int ii;
 
 	if (index >= (8 * sizeof(n)))
